Remove suppression code for false vulnerabilities
Several blocks of suppression code have been removed from the project-suppression.xml file due to false or resolved vulnerabilities. The suppression blocks that have been removed were found to be unnecessary as they were linked to vulnerabilities that were proven to be non-issues or already fixed. This simplifies the suppression file and reduces room for confusion.